feat: move executor types out into util.
